:truck: Rename ttt_game/ to game/
