+++ src/game/ttt_game.h
@@ -23,21 +23,24 @@
  */
 
  /*
-     Name : ttt_game/ttt_input.h
+     Name : game/ttt_game.h
      Author : Antoine James Tournepiche
      Creation Date : December 24th 2019
      Last update : December 24th 2019
      Project : ASCII Tic Tac Toe
      Project sources : https://github.com/AntoineJT/ascii-tic-tac-toe
 
-     TicTacToe input header
+     TicTacToe game logic header
  */
 
-#ifndef _H_TTT_INPUT_
-#define _H_TTT_INPUT_
+#ifndef _H_TTT_GAME_
+#define _H_TTT_GAME_
 
+#include "../boolean.h"
 #include "ttt_board.h"
 #include "ttt_player.h"
 
-void ttt_input_cell(ttt_board* board, const ttt_player player, unsigned int* cell);
+bool ttt_is_cell_line_claimed_by_one_player(ttt_board board, const ttt_player a, const ttt_player b, const ttt_player c);
+bool ttt_play_cell(ttt_board* board, const unsigned int cell, const ttt_player player);
+ttt_player ttt_get_winner(const ttt_board board);
 #endif
