+++ src/game/ttt_player.c
@@ -22,29 +22,38 @@
  *  SOFTWARE.
  */
 
-/*
-    Name : ttt_game/ttt_player.h
-    Author : Antoine James Tournepiche
-    Creation Date : December 20th 2019
-    Last update : December 24th 2019
-    Project : ASCII Tic Tac Toe
-    Project sources : https://github.com/AntoineJT/ascii-tic-tac-toe
-    
-    TicTacToe player header
-*/
+ /*
+     Name : game/ttt_player.c
+     Author : Antoine James Tournepiche
+     Creation Date : December 24th 2019
+     Last update : December 24th 2019
+     Project : ASCII Tic Tac Toe
+     Project sources : https://github.com/AntoineJT/ascii-tic-tac-toe
 
-#ifndef _H_TTT_PLAYER_
-#define _H_TTT_PLAYER_
+     TicTacToe player source code file
+ */
 
-typedef enum
+#include "ttt_player.h"
+
+char* ttt_get_player_name(const ttt_player player)
 {
-    PLAYER_CROSS,
-    PLAYER_CIRCLE,
-    PLAYER_NULL,
-    // Used to indicates when no one get a frame or when it has no winner for now
-    PLAYER_UNDEFINED // Used to indicates when it has equality
-} ttt_player;
+    switch (player)
+    {
+    case PLAYER_CROSS: return "Cross";
+    case PLAYER_CIRCLE: return "Circle";
+    default: return "ERROR";
+    }
+}
 
-char* ttt_get_player_name(const ttt_player player);
-ttt_player ttt_get_opponent(const ttt_player player);
-#endif
+ttt_player ttt_get_opponent(const ttt_player player)
+{
+    switch (player)
+    {
+    case PLAYER_CIRCLE: return PLAYER_CROSS;
+    case PLAYER_CROSS: return PLAYER_CIRCLE;
+    case PLAYER_NULL:
+    case PLAYER_UNDEFINED:
+    default:
+        return PLAYER_NULL;
+    }
+}
